sft/requirements.txt
absl-py==2.3.1
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
airportsdata==20250909
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
astor==0.8.1
asttokens==3.0.0
attrs==25.4.0
audioread==3.1.0
av==16.0.1
bitsandbytes==0.48.2
blake3==1.0.8
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
compressed-tensors==0.9.3
contourpy==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
deepspeed==0.16.9
Deprecated==1.3.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.121.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
ffmpy==0.6.4
filelock==3.20.0
fire==0.7.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.3.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.71.0
gradio==5.31.0
gradio_client==1.10.1
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.0.0
interegular==0.3.3
ipdb==0.13.13
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jax==0.8.1
jaxlib==0.8.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
librosa==0.11.0
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral_common==1.8.5
ml_dtypes==0.5.4
modelscope==1.31.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==2.7.1
opencv-python-headless==4.12.0.88
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.13
opt_einsum==3.4.0
orjson==3.11.4
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.3
parso==0.8.5
partial-json-parser==0.2.1.1.post6
peft==0.15.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pycountry==24.6.1
pycparser==2.23
pydantic==2.10.6
pydantic-extra-types==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
ray==2.51.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rouge-chinese==1.0.3
rpds-py==0.28.0
ruff==0.14.3
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.43.0
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
triton==3.2.0
trl==0.9.6
typer==0.20.0
typing_extensions==4.15.0
tyro==0.8.14
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.8.5.post1
wandb==0.23.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.0
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
